+++ webpack.config.js
@@ -1,0 +1,31 @@
+const path = require('path')
+const webpack = require('webpack')
+
+module.exports = {
+  target: 'node',
+  plugins: [
+    new webpack.BannerPlugin({
+      banner: '#!/usr/bin/env node',
+      raw: true
+    })
+  ],
+  // dependencies externals
+  externals: ['aws-sdk', 'electron'],
+  entry: './src/index.ts',
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/
+      }
+    ]
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js']
+  },
+  output: {
+    filename: 'index.js',
+    path: path.resolve(__dirname, 'dist', 'cli')
+  }
+}
